refactor(scan-nh): split various functions into their own files
This is an attempt to make CI stop running out of memory.
