+++ libs/converter-nh-accuvote/src/convert/read_grid_from_election_definition.test.ts
@@ -1,0 +1,96 @@
+import { electionGridLayoutNewHampshireHudsonFixtures } from '@votingworks/fixtures';
+import { readFixtureDefinition } from '../../test/fixtures';
+import { readGridFromElectionDefinition } from './read_grid_from_election_definition';
+import { asciiOvalGrid } from '../../test/utils';
+
+test('readGridFromElectionDefinition', () => {
+  const definition = readFixtureDefinition(
+    electionGridLayoutNewHampshireHudsonFixtures.definitionXml.asText()
+  );
+  const grid = readGridFromElectionDefinition(definition);
+  expect(asciiOvalGrid(grid)).toMatchInlineSnapshot(`
+    "                                  
+                                      
+                                      
+                                      
+                                      
+                                      
+                                      
+                                      
+                                      
+                O      O      O       
+                                      
+                                     O
+                                      
+                                      
+                                      
+                O      O      O      O
+                                      
+                                      
+                                      
+                O      O      O      O
+                                      
+                                      
+                                      
+                O      O      O       
+                                      
+                                     O
+                                      
+                                      
+                                      
+                O      O             O
+                                      
+                                      
+                                      
+                O      O             O
+                                      
+                                      
+                                      
+                O      O             O
+                                      
+                O      O             O
+                                      
+                O      O             O
+                                      
+                O      O             O
+                                      
+                O      O             O
+                                      
+                O      O             O
+                                      
+                O      O             O
+                                      
+                O      O             O
+                                      
+                O      O             O
+                                      
+                O      O             O
+                                      
+                O      O             O
+                                      
+                                      
+                                      
+                O      O             O
+                                      
+                                      
+                                      
+                O      O      O      O
+                                      
+                                      
+                                      
+                O      O      O      O
+                                      
+                                      
+                                      
+                O      O             O
+                                      
+                                      
+                                      
+                O      O             O
+                                      
+                                      
+                                      
+                O      O             O
+    "
+  `);
+});
