+++ libs/converter-nh-accuvote/src/index.ts
@@ -1,9 +1,6 @@
 /* istanbul ignore file - this file only exists as the library entry-point–it shouldn't have any code to test */
 
-export {
-  convertElectionDefinition,
-  ConvertIssueKind,
-  type ConvertIssue,
-} from './convert';
+export { convertElectionDefinition } from './convert/convert_election_definition';
+export { ConvertIssueKind, type ConvertIssue } from './convert/types';
 export * as templates from './data/templates';
 export * from './layout';
